remove usedIVs key from VaultItem, various fixes/formatting.
breaking changes from previous commit.

prepare vault backups to have a sequenced nonce that we can track.  Remove outer mac on VaultItem, should use the derived key for this but may not have in some instances (ie. using the pin code derived keys to decrypt).  Originally was going to be a double mac check on outer metadata, may re-implement but removed for now.

standardize mac computation on NoteItem and PasswordItem objects.